Add confirmation for canceling reservation

+++ HetDepot/Controllers/ReservationRemoveController.cs
@@ -1,6 +1,7 @@
 ﻿using HetDepot.People.Model;
 using HetDepot.Tours.Model;
 using HetDepot.Views;
+using HetDepot.Views.Parts;
 
 namespace HetDepot.Controllers
 {
@@ -17,13 +18,35 @@ namespace HetDepot.Controllers
 
 		public override void Execute()
 		{
-			Program.TourService.RemoveTourReservation(_tour, _visitor);
-
-			var message = Program.SettingService.GetConsoleText("consoleVisitorReservationCancellationConfirmation");
-
-			new AlertView(message, AlertView.Info).Show();
-
 			NextController = new ShowToursController();
+
+			var question = Program.SettingService.GetConsoleText("consoleVisitorReservationAlreadyHavingOne");
+			
+			ListView replaceCurrentQuestion = new ListView(question, new List<ListableItem>()
+			{
+				new ListViewItem("Ja", true),
+				new ListViewItem("Nee", false),
+			});
+			bool cancelReservartion = (bool)replaceCurrentQuestion.ShowAndGetResult();
+
+			if (!cancelReservartion)
+			{
+				var messageNotCancled = Program.SettingService
+					.GetConsoleText("consoleVisitorReservationChangeNotChanges", new()
+					{
+						["time"] = _tour.GetTime(),
+					});
+				
+				new AlertView(messageNotCancled, AlertView.Info).Show();
+
+				return;
+			}
+			
+			Program.TourService.RemoveTourReservation(_tour, _visitor);
+			
+			var messageCancled = Program.SettingService.GetConsoleText("consoleVisitorReservationCancellationConfirmation");
+
+			new AlertView(messageCancled, AlertView.Info).Show();
 		}
 	}
 }
